+++ changelog.md
@@ -1,5 +1,6 @@
 ### 0.1.5 - 2018-09-27
 - Improved log statements for better debugging
+- Any type of `error` event will be emitted with a tick dealy. This would give enough time for the `create()` methods to resolve the promise.
 
 ### 0.1.4 - 2018-09-25
 - `options` is a required property of `Connection` and `Container`.